+++ common/src/main/java/smartin/miapi/item/modular/items/ExampleModularStrackableItem.java
@@ -23,6 +23,10 @@ public class ExampleModularStrackableItem extends Item implements ModularItem {
         modularItem = this;
     }
 
+    public ExampleModularStrackableItem(Item.Settings settings) {
+        super(settings);
+    }
+
     @Override
     public Text getName(ItemStack stack) {
         return DisplayNameProperty.getDisplayText(stack);