+++ forge/src/main/java/smartin/miapi/forge/MiapiReloadListenerForge.java
@@ -3,7 +3,7 @@ package smartin.miapi.forge;
 import com.google.gson.JsonObject;
 import net.minecraft.resource.Resource;
 import net.minecraft.resource.ResourceManager;
-import net.minecraft.resource.ResourceReloader;
+import net.minecraft.resource.SinglePreparationResourceReloader;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.profiler.Profiler;
@@ -13,15 +13,60 @@ import smartin.miapi.modules.conditions.ConditionManager;
 
 import java.io.BufferedReader;
 import java.util.*;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Executor;
 import java.util.stream.Collectors;
 
 
-public class MiapiReloadListenerForge implements ResourceReloader {
+public class MiapiReloadListenerForge extends SinglePreparationResourceReloader<Map<String,String>> {
     static long timeStart;
 
-    public CompletableFuture load(ResourceManager manager, Profiler profiler, Executor executor) {
+    protected void apply(Map<String, String> prepared, ResourceManager manager, Profiler profiler) {
+        Map < String, String> dataMap = new HashMap<>((Map) prepared);
+        Map<String, String> filteredMap = new HashMap<>();
+        dataMap.forEach((key, value) -> {
+            if (!key.endsWith(".json")) {
+                filteredMap.put(key, value);
+                return;
+            }
+            try {
+                JsonObject element = Miapi.gson.fromJson(value, JsonObject.class);
+                if (!element.has("load_condition")) {
+                    filteredMap.put(key, value);
+                    return;
+                }
+                boolean allowed = ConditionManager.get(element.get("load_condition")).isAllowed(new ConditionManager.ConditionContext() {
+                    @Override
+                    public ConditionManager.ConditionContext copy() {
+                        return this;
+                    }
+
+                    @Override
+                    public List<Text> getReasons() {
+                        return new ArrayList<>();
+                    }
+                });
+                if (allowed) {
+                    element.remove("load_condition");
+                    Miapi.LOGGER.info("redid " + key);
+                    filteredMap.put(key, Miapi.gson.toJson(element));
+                }
+            } catch (Exception e) {
+                filteredMap.put(key, value);
+            }
+        });
+
+
+        ReloadEvents.DataPackLoader.trigger(filteredMap);
+        ReloadEvents.MAIN.fireEvent(false);
+        ReloadEvents.END.fireEvent(false);
+        Miapi.LOGGER.info("Server load took " + (double) (System.nanoTime() - timeStart) / 1000 / 1000 + " ms");
+        ReloadEvents.reloadCounter--;
+        if (Miapi.server != null) {
+            Miapi.server.getPlayerManager().getPlayerList().forEach(ReloadEvents::triggerReloadOnClient);
+        }
+    }
+
+    @Override
+    protected Map<String, String> prepare(ResourceManager manager, Profiler profiler) {
         ReloadEvents.reloadCounter++;
         timeStart = System.nanoTime();
         ReloadEvents.START.fireEvent(false);
@@ -46,59 +91,6 @@ public class MiapiReloadListenerForge implements ResourceReloader {
                 });
             });
         });
-        return CompletableFuture.completedFuture(data);
-    }
-
-    public CompletableFuture<Void> apply(Object data, ResourceManager manager, Profiler profiler, Executor executor) {
-        return CompletableFuture.runAsync(() -> {
-            Map<String, String> dataMap = new HashMap<>((Map) data);
-            Map<String, String> filteredMap = new HashMap<>();
-            dataMap.forEach((key, value) -> {
-                if (!key.endsWith(".json")) {
-                    filteredMap.put(key, value);
-                    return;
-                }
-                try {
-                    JsonObject element = Miapi.gson.fromJson(value, JsonObject.class);
-                    if (!element.has("load_condition")) {
-                        filteredMap.put(key, value);
-                        return;
-                    }
-                    boolean allowed = ConditionManager.get(element.get("load_condition")).isAllowed(new ConditionManager.ConditionContext() {
-                        @Override
-                        public ConditionManager.ConditionContext copy() {
-                            return this;
-                        }
-
-                        @Override
-                        public List<Text> getReasons() {
-                            return new ArrayList<>();
-                        }
-                    });
-                    if (allowed) {
-                        element.remove("load_condition");
-                        Miapi.LOGGER.info("redid " + key);
-                        filteredMap.put(key, Miapi.gson.toJson(element));
-                    }
-                } catch (Exception e) {
-                    filteredMap.put(key, value);
-                }
-            });
-
-
-            ReloadEvents.DataPackLoader.trigger(filteredMap);
-            Miapi.LOGGER.info("Server load took " + (double) (System.nanoTime() - timeStart) / 1000 / 1000 + " ms");
-            if (Miapi.server != null) {
-                Miapi.server.getPlayerManager().getPlayerList().forEach(ReloadEvents::triggerReloadOnClient);
-            }
-            ReloadEvents.reloadCounter--;
-        });
-    }
-
-    @Override
-    public CompletableFuture<Void> reload(Synchronizer synchronizer, ResourceManager manager, Profiler prepareProfiler, Profiler applyProfiler, Executor prepareExecutor, Executor applyExecutor) {
-        return load(manager, prepareProfiler, prepareExecutor).thenCompose(synchronizer::whenPrepared).thenAcceptAsync(a -> {
-            apply(a, manager, applyProfiler, applyExecutor);
-        });
+        return data;
     }
 }