+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 - fixed beeing unable to pickup arrows
 - fixed making arrows non stackable sometimes
 - added new /miapi clear_cache and /miapi miapi_reload to help debugging
+- adjusted reload logic on forge for better compatibility
+- fixed wood and stone materials breaking
+- adjusted double reload logic for stability